Ajout du homeController, d'un lien, et retirer commentaires
