+++ views/index.ejs
@@ -1,42 +1,47 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Index</title>
     <link rel="stylesheet" href="css/styles.css">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet"
+        integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
 </head>
+
 <body class="flex-container-between">
     <header>
         <nav>
-            <div class="logo"><img src="/img/logo-p.png"></div>
-           <div class="main_menu">
-               <a href="/"><img src="/img/home.png" alt="home"></a>
-               <a href="#"><img src="/img/chat.png" alt="chat"></a>
-               <a href="/profile"><img src="/img/profil.png" alt="profil"></a>
-               <a href="#"><img src="/img/spot.png" alt="spot"></a>
-           </div>
-           <form class="flex-container">
-            <input class="me-2 search-nav" type="search" placeholder="Search" aria-label="Search">
-            <button class="btn" type="submit"><img src="/img/search.png" alt="search"></button>
-          </form>
-     </nav>
+            <div class="logo"><a href="/"><img src="/img/logo-p.png"></a></div>
+            <div class="main_menu">
+                <a href="/"><img src="/img/home.png" alt="home"></a>
+                <a href="#"><img src="/img/chat.png" alt="chat"></a>
+                <a href="/profile"><img src="/img/profil.png" alt="profil"></a>
+                <a href="#"><img src="/img/spot.png" alt="spot"></a>
+            </div>
+            <form class="flex-container">
+                <input class="me-2 search-nav" type="search" placeholder="Search" aria-label="Search">
+                <button class="btn" type="submit"><img src="/img/search.png" alt="search"></button>
+            </form>
+        </nav>
     </header>
     <main>
         <div class="container flex-container">
             <div class="logo-container">
-            <img src="/img/logo.png" alt="Logo">
-        </div>
+                <img src="/img/logo.png" alt="Logo">
+            </div>
         </div>
         <div class="container">
             <div class="row">
                 <div class="col-12">
                     <div class="presentation">
-                        Tu as de la misère à trouver de la poudreuse qui n’a pas été tapée par un millier de skieurs de randonnée ?<br>
+                        Tu as de la misère à trouver de la poudreuse qui n’a pas été tapée par un millier de skieurs de
+                        randonnée ?<br>
                         <span>Tu es au bon endroit !</span><br>
-                        Vient découvrir les meilleurs spots et fait découvrir tes secrets bien gardés... un partage à la fois !
+                        Vient découvrir les meilleurs spots et fait découvrir tes secrets bien gardés... un partage à la
+                        fois !
                     </div>
                 </div>
             </div>
@@ -51,7 +56,7 @@
                             Ou de partager les vôtres ?<br> <br>
 
                             Rejoint-nous!</div>
-                            <span><a href="/signup"><button type="submit">S'inscrire</button></a>
+                        <span><a href="/signup"><button type="submit">S'inscrire</button></a>
                     </div>
                 </div>
                 <div class="col-6 flex-container">
@@ -63,7 +68,7 @@
 
                             C’est le moment de découvrir ce que tes amis<br>
                             ont fait durant ton absence!</div>
-                            <a href="/signin"><button type="submit">Connexion</button></a>
+                        <a href="/signin"><button type="submit">Connexion</button></a>
                     </div>
                 </div>
             </div>
@@ -71,8 +76,9 @@
     </main>
     <footer>
         <a href="/"><img src="/img/logo-p.png"></a>
-        <div>Termes et Conditions  -  Confidentialité  -  À propos  -  Nous contacter</div>
+        <div>Termes et Conditions - Confidentialité - À propos - Nous contacter</div>
         <div><img src="/img/socialSkieur.png" alt=""></div>
     </footer>
 </body>
+
 </html>
